Add snapcraft.yaml defining the snap package and its components
This initial package comes with a number of constraints:

  * LDC uses `gcc` for linking, and the D standard library needs to link
    against the C standard library.  Since a containerized snap cannot
    access host-system `gcc` and `libc` (and indeed on an Ubuntu Core
    system can't even assume `gcc` is available!), Ubuntu `gcc`, `libc6`
    and `libc6-dev` packages are bundled in as part of the snap, with a
    GCC wrapper script used to ensure the sysroot path is set correctly.
    In future this can probably be replaced by a dependency on GCC and
    libc interfaces.

  * Somewhere in the intersection of how snapcraft specifies the install
    path, and LDC's build process, the autogenerated `ldc2.conf` winds
    up with incorrect include and library paths.  Without any (apparent)
    means to ask LDC's build process to generate a config file that uses
    relative path definitions, for now a correct `ldc2.conf` is provided
    in the snap package, and copied to the staging and priming areas in
    preference to the file auto-generated by the LDC build.

However, the package as a whole should provide a reliably self-contained
LDC, exposing `ldc2`, `ldmd2` and `ldc-profdata` (note that the latter
two of these are callable only as `ldc2.ldmd2` and `ldc2.ldc-profdata`).

Note that `ldc` itself is required as a build dependency, since the LDC
build process requires a working D compiler to compile the DMD frontend
(itself written in D).  On an Ubuntu 16.04 system, LDC 0.17.1 will be
used (the last LDC version to use the old C++ DMD frontend).